Extract Open Board import hook
